pod-eviction-controller/requirements.txt
kubernetes~=33.1.0
kopf~=1.38.0
python-dateutil~=2.9.0.post0